feat: update Normalize(), added TrimAndRemoveZeroWidthChars()
