+++ string.go
@@ -1,7 +1,6 @@
 package agstring
 
 import (
-	"fmt"
 	"regexp"
 	"strconv"
 	"strings"
@@ -24,107 +23,6 @@ func First(ls ...string) string { return Nth(ls, 0) }
 
 // Last returns the last element of given list or empty string when the list is empty.
 func Last(ls ...string) string { return Nth(ls, len(ls)-1) }
-
-// TrimSuffixes returns s without any of the provided trailing suffixes strings.
-func TrimSuffixes(s string, suffixes ...string) string {
-	s = strings.TrimSpace(s)
-	for _, suffix := range suffixes {
-		if strings.HasSuffix(s, suffix) {
-			return strings.TrimSpace(strings.TrimSuffix(s, suffix))
-		}
-	}
-	return s
-}
-
-// TrimAllSuffixes returns a string without any of the provided trailing suffixes or spaces.
-// See test for examples.
-func TrimAllSuffixes(s string, suffixes ...string) string {
-	if len(suffixes) == 0 || s == "" {
-		return strings.TrimSpace(s)
-	}
-
-	reSufs := make([]*regexp.Regexp, 0)
-	for _, suffix := range suffixes {
-		if suffix == "" {
-			continue
-		}
-
-		reE := fmt.Sprintf(`^(?P<rest>.*)%s\s*$`, regexp.QuoteMeta(suffix))
-		reSufs = append(reSufs, regexp.MustCompile(reE))
-	}
-
-	trimAgain := true
-	for trimAgain {
-		trimAgain = false
-		for _, reSuf := range reSufs {
-			if matches, ok := RegexpGroups(reSuf, strings.TrimSpace(s)); ok {
-				s = matches["rest"]
-				trimAgain = true
-			}
-		}
-	}
-	return strings.TrimSpace(s)
-}
-
-// TrimAllPrefixes returns a string without any of the provided leading prefixes or spaces.
-// See test for examples.
-func TrimAllPrefixes(s string, prefixes ...string) string {
-	if len(prefixes) == 0 || s == "" {
-		return strings.TrimSpace(s)
-	}
-
-	rePres := make([]*regexp.Regexp, 0)
-	for _, prefix := range prefixes {
-		if prefix == "" {
-			continue
-		}
-
-		reE := fmt.Sprintf(`^\s*%s(?P<rest>.*)`, regexp.QuoteMeta(prefix))
-		rePres = append(rePres, regexp.MustCompile(reE))
-	}
-
-	trimAgain := true
-	for trimAgain {
-		trimAgain = false
-		for _, rePre := range rePres {
-			if matches, ok := RegexpGroups(rePre, strings.TrimSpace(s)); ok {
-				s = matches["rest"]
-				trimAgain = true
-			}
-		}
-	}
-	return strings.TrimSpace(s)
-}
-
-// TrimPrefixesAndSpace returns a string without any of the provided leading prefixes at word
-// boundaries or spaces. See test for examples.
-func TrimPrefixesAndSpace(s string, prefixes ...string) string {
-	if prefixes == nil || s == "" {
-		return s
-	}
-
-	rePres := make([]*regexp.Regexp, 0)
-	for _, prefix := range prefixes {
-		if prefix == "" {
-			continue
-		}
-
-		reE := fmt.Sprintf("^\\s*%s\\b(?P<rest>.*)", regexp.QuoteMeta(prefix))
-		rePres = append(rePres, regexp.MustCompile(reE))
-	}
-
-	trimAgain := true
-	for trimAgain {
-		trimAgain = false
-		for _, rePre := range rePres {
-			if matches, ok := RegexpGroups(rePre, strings.TrimSpace(s)); ok {
-				s = matches["rest"]
-				trimAgain = true
-			}
-		}
-	}
-	return strings.TrimSpace(s)
-}
 
 var nonAlphanumRegexp = regexp.MustCompile("[^[:alnum:]]")
 
@@ -168,9 +66,6 @@ type StringIterator interface {
 	Get() string
 	HasNext() bool
 }
-
-// TrimSpace trims spaces in the given slice
-func TrimSpace(ls ...string) []string { return Map(ls, strings.TrimSpace) }
 
 // ToLower makes lowercase strings in the given slice
 func ToLower(ls ...string) []string { return Map(ls, strings.ToLower) }
